exec-engine/requirements.txt
fastapi
uvicorn
playwright
openai
python-dotenv
pyjwt
pydantic
sqlalchemy
pymysql
faker
httpx
